refactor: Rename functions and variables, change function arguments
The variable element_type is now called tag_type. 

The function to convert a line to HTML is renamed from convert_line_to_html, which takes an unparsed line as input, to convert_tag_to_html, which takes a parsed line as input. This way, the line need not be parsed twice.